+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.2.0]
+
+* `controller` argument in `ParallaxImage` constructor is no longer required.
+  If controller is not specified then `ParallaxImage` looks for nearest
+  `Scrollable` ancestor and subscribes to scrolling updates on it.
+
 ## [0.1.0]
 
 * Initial release